Move manifest compiler/decompiler to a module
