+++ simulator/src/resim/mod.rs
@@ -55,7 +55,7 @@ use scrypto::crypto::*;
 use std::env;
 use std::fs;
 use std::path::PathBuf;
-use transaction_manifest::decompile;
+use transaction::manifest::decompile;
 
 use crate::ledger::*;
